pyproject.toml
[project]
name = "downloader-tg-py"
version = "0.1.0"
description = "Bot for downloading files via links in tg"
readme = "readme.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "pyrofork>=2.3.61",
    "tgcrypto-pyrofork>=1.2.7",
    "tortoise-orm>=0.24.2",
    "yt-dlp>=2025.3.31",
]

[tool.pyright]
venvPath = '.'
venv = '.venv'
